Refactor project directory structure according to task.
